# Improve `MessageMapping` rules dispatch