--- a/.gitignore
+++ b/.gitignore
@@ -35,3 +35,5 @@ examples/*Compiled
 
 **/.benchmarkci
 benchmark/*.json
+
+Session.vim