--- a/src/message.jl
+++ b/src/message.jl
@@ -270,6 +270,7 @@ end
 ## However it is not fully inferrable due to dynamic tags and variable constraints, but still better than just a raw lambda callback
 
 struct MessageMapping{F, T, C, N, M, A, X, R}
+    rulefn          :: F
     vtag            :: T
     vconstraint     :: C
     msgs_names      :: N
@@ -278,9 +279,6 @@ struct MessageMapping{F, T, C, N, M, A, X, R}
     addons          :: X
     factornode      :: R
 end
-
-message_mapping_fform(::MessageMapping{F}) where {F} = F
-message_mapping_fform(::MessageMapping{F}) where {F <: Function} = F.instance
 
 # Some addons add post rule execution logic
 function message_mapping_addons(mapping::MessageMapping, messages, marginals, result, addons)
@@ -305,14 +303,6 @@ end
 # Other addons may override this behaviour (if necessary, see e.g. AddonMemory)
 message_mapping_addon(addon, mapping, messages, marginals, result) = addon
 
-function MessageMapping(::Type{F}, vtag::T, vconstraint::C, msgs_names::N, marginals_names::M, meta::A, addons::X, factornode::R) where {F, T, C, N, M, A, X, R}
-    return MessageMapping{F, T, C, N, M, A, X, R}(vtag, vconstraint, msgs_names, marginals_names, meta, addons, factornode)
-end
-
-function MessageMapping(::F, vtag::T, vconstraint::C, msgs_names::N, marginals_names::M, meta::A, addons::X, factornode::R) where {F <: Function, T, C, N, M, A, X, R}
-    return MessageMapping{F, T, C, N, M, A, X, R}(vtag, vconstraint, msgs_names, marginals_names, meta, addons, factornode)
-end
-
 function materialize!(mapping::MessageMapping, dependencies)
     messages  = dependencies[1]
     marginals = dependencies[2]
@@ -323,8 +313,7 @@ function materialize!(mapping::MessageMapping, dependencies)
     # Message is initial if it is not clamped and all of the inputs are either clamped or initial
     is_message_initial = !is_message_clamped && (__check_all(is_clamped_or_initial, messages) && __check_all(is_clamped_or_initial, marginals))
 
-    result, addons = rule(
-        message_mapping_fform(mapping),
+    result, addons = mapping.rulefn(
         mapping.vtag,
         mapping.vconstraint,
         mapping.msgs_names,