--- a/src/nodes/delta/delta.jl
+++ b/src/nodes/delta/delta.jl
@@ -69,9 +69,11 @@ nodefunction(factornode::DeltaFnNode, ::Val{:in})             = getinverse(metad
 nodefunction(factornode::DeltaFnNode, ::Val{:in}, k::Integer) = getinverse(metadata(factornode), k)
 
 # Rules for `::Function` objects, but with the `DeltaFn` related meta and node should redirect to the `DeltaFn` rules
-function rule(::F, on, vconstraint, mnames, messages, qnames, marginals, meta::DeltaMeta, addons::Any, node::DeltaFnNode) where {F <: Function}
-    return rule(DeltaFn{F}, on, vconstraint, mnames, messages, qnames, marginals, meta, addons, node)
-end
+# function rule(::F, on, vconstraint, mnames, messages, qnames, marginals, meta::DeltaMeta, addons::Any, node::DeltaFnNode) where {F <: Function}
+    # return rule(DeltaFn{F}, on, vconstraint, mnames, messages, qnames, marginals, meta, addons, node)
+# end
+# NOTE: dispatch for `rule` is optimized by hand, the method above is no longer needed, but `marginalrule` are called 
+# in less efficient way, remove the `marginalrule` below when it is optimized in the same way as for the `rule`
 
 function marginalrule(::F, on, mnames, messages, qnames, marginals, meta::DeltaMeta, node::DeltaFnNode) where {F <: Function}
     return marginalrule(DeltaFn{F}, on, mnames, messages, qnames, marginals, meta, node)