--- a/src/nodes/implication.jl
+++ b/src/nodes/implication.jl
@@ -1,7 +1,7 @@
 export IMPLY
 
 """
-IMPY node implements implication function that can be desribed by the followsing table:
+IMPLY node implements implication function that can be desribed by the followsing table:
 | in1  in2 | out |
 |  0    0  |  1  |
 |  0    1  |  1  |