--- a/src/rules/implication/in2.jl
+++ b/src/rules/implication/in2.jl
@@ -1,4 +1,4 @@
-@rule typeof(IMPLY)(:in2, Marginalisation) (m_out::Bernoulli, m_in1::Bernoulli) = begin
+@rule IMPLY(:in2, Marginalisation) (m_out::Bernoulli, m_in1::Bernoulli) = begin
     pout, pin1 = mean(m_out), mean(m_in1)
 
     return Bernoulli((pout) / (2 * pout + pin1 - 2 * pout * pin1))