This PR improves upon #290 and addresses #284 even further. 

This PR changes the way `MessageMapping` dispatches on a particular rule. I couldn't improve things further because it requires a lot of changes in the core, but there is definitely room for improvement and we can reduce allocation profile even further. This is very crucial for real-time applications.

The PR is breaking, but imo minor. After this PR, it is no longer possible to add/change/override built-in rules without explicit import statement, e.g:
```
import ReactiveMP: rule_for_NormalMeanVariance # `@rule` below will error without this import statement

@rule NormalMeanVariance(:out, ...) ... = begin 
 ...
end
```
There are no problems adding new rules with the explicit import statement. I could think more of possible solutions for that, but in general further improvements require a complete `rule` refactoring.

The benefits are improved allocation profile and faster inference in general, especially for the `rxinference` function from `RxInfer`. These are timing from the `Tiny Benchmark.ipynb`. 

### `master`

```
BenchmarkTools.Trial: 1834 samples with 1 evaluation.
 Range (min … max):  2.457 ms … 11.236 ms  ┊ GC (min … max): 0.00% … 74.35%
 Time  (median):     2.518 ms              ┊ GC (median):    0.00%
 Time  (mean ± σ):   2.725 ms ±  1.177 ms  ┊ GC (mean ± σ):  6.67% ± 11.31%

  █▅▁                                                         
  ███▆▄▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▅▆▇ █
  2.46 ms      Histogram: log(frequency) by time     10.1 ms <

 Memory estimate: 3.05 MiB, allocs estimate: 83828.
BenchmarkTools.Trial: 565 samples with 1 evaluation.
 Range (min … max):  7.731 ms … 21.617 ms  ┊ GC (min … max): 0.00% … 54.27%
 Time  (median):     7.899 ms              ┊ GC (median):    0.00%
 Time  (mean ± σ):   8.850 ms ±  2.809 ms  ┊ GC (mean ± σ):  9.13% ± 14.68%

  █▆▂▂                                                        
  █████▆▅▆▆▅▄▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▄▄▆▆▁▄▅▆▇▆▅▇▅▄ ▆
  7.73 ms      Histogram: log(frequency) by time     18.8 ms <

 Memory estimate: 11.10 MiB, allocs estimate: 241913.
```

### `PR`

```
BenchmarkTools.Trial: 2831 samples with 1 evaluation.
 Range (min … max):  1.576 ms …   9.150 ms  ┊ GC (min … max): 0.00% … 79.29%
 Time  (median):     1.615 ms               ┊ GC (median):    0.00%
 Time  (mean ± σ):   1.765 ms ± 998.268 μs  ┊ GC (mean ± σ):  8.03% ± 11.32%

  █▂                                                           
  ██▅▃▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁█ █
  1.58 ms      Histogram: log(frequency) by time      8.65 ms <

 Memory estimate: 2.56 MiB, allocs estimate: 67828.

BenchmarkTools.Trial: 603 samples with 1 evaluation.
 Range (min … max):  6.953 ms … 21.804 ms  ┊ GC (min … max):  0.00% … 55.96%
 Time  (median):     7.264 ms              ┊ GC (median):     0.00%
 Time  (mean ± σ):   8.295 ms ±  3.085 ms  ┊ GC (mean ± σ):  10.30% ± 15.45%

  █▆▅▃▂▁ ▁                                                    
  ████████▆▅▆▅▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▄▁▁▄▄▆▇██▆▆▅▅▁▄▄ ▇
  6.95 ms      Histogram: log(frequency) by time     19.8 ms <

 Memory estimate: 10.73 MiB, allocs estimate: 229906.
```

WDYT of the benefits and of the breaking change? 

@bartvanerp you might be interested.