@bvdmitri 2023-03-03
This is the README benchmark with this PR
![image](https://user-images.githubusercontent.com/6557701/222767548-6df6810a-bf76-4b44-a2df-75c150c68e70.png)